+++ src/01-inbound/port/i.service.ts
@@ -1,0 +1,14 @@
+import { IArticleCommentService } from "./services/i.article.comment.service";
+import { IArticleService } from "./services/i.article.service";
+import { IProductCommentService } from "./services/i.product.comment.service";
+import { IProductService } from "./services/i.product.service";
+import { IUserService } from "./services/i.user.service";
+
+export interface IService {
+    productService: IProductService,
+    articleService: IArticleService,
+    userService: IUserService,
+    articleCommentService:  IArticleCommentService,
+    productCommentService: IProductCommentService
+}
+
