+++ src/01-inbound/response/article.res.dto.ts
