+++ src/01-inbound/response/product.res.dto.ts
