--- src/02-controller/req-validator/req.validator.ts
@@ -1,105 +1,0 @@
-import { z } from "zod";
-
-export const querySchema = z.object({
-    offset: z.coerce.number({ message: "offset이 유효하지 않습니다" }).default(0).optional(),
-    limit: z.coerce.number().default(10).optional(),
-    search: z.string().optional().default("").optional(),
-    sort: z.string().optional().default("desc").optional()
-})
-
-
-
-export const articleBodySchema = querySchema.extend({
-    title: z.string(),
-    content: z.string(),
-})
-
-export const articleParamSchema = z.object({
-    id: z.string().optional()
-})
-
-
-
-
-
-
-
-
-
-
-
-export const productReqSchema = querySchema.extend({
-    body: z.object({
-        name: z.string(),
-        description: z.string(),
-        price: z.coerce.number(),
-        tags: z.array(z.string()).optional(),
-        imageUrl: z.string().optional(),
-        isLiked: z.boolean().default(false)
-    }),
-
-    user: z.object({
-        userId: z.string()
-    }),
-
-    params: z.object({
-        id: z.string().optional()
-    }).optional()
-})
-
-export const userReqSchema = z.object({
-    email: z.string(),
-    nickname: z.string(),
-    password: z.string(),
-
-})
-
-
-
-
-export const productCommentBodySchema  = z.object({
-    content: z.string()
-})
-
-export const productCommentParamSchema = z.object({
-    productId: z.string(),
-    commentId: z.string().optional()
-})
-
-
-
-
-export const articleCommentBodySchema = z.object({
-    content: z.string()
-})
-
-export const articleCommentParamSchema = z.object({
-    articleId: z.string(),
-    commentId: z.string().optional()
-})
-
-
-
-
-export type QueryType = z.infer<typeof querySchema>;
-
-export type ArticleReqDto = z.infer<typeof articleBodySchema> &
-    z.infer<typeof articleParamSchema> & {
-        userId: string
-    };
-
-export type ProductRequest = z.infer<typeof productReqSchema>;
-
-export type UserSignUpDto = z.infer<typeof userReqSchema>;
-export type UserSignInDto = Omit<z.infer<typeof userReqSchema>, 'nickname'>;
-
-
-export type ProductCommentRequest = z.infer<typeof  productCommentBodySchema> &
-    z.infer<typeof productCommentParamSchema> & {
-        userId: string
-    };
-export type ArticleCommentRequest = z.infer<typeof articleCommentBodySchema> &
-    z.infer<typeof articleCommentParamSchema> & {
-        userId: string
-    };
-
